Summary Project
The present Project had been completed in the period of 3 months from March 2024 to May 2024. The data was cleaned using power query editor. The data analysis was performed by answering some questions related to the key metrics of Zomato through SQL and further some important charts and dashboards were prepared in Excel, Power Bi and Tableau. 
On the completion of the projects the findings of the project suggested that Zomato need to improve its marketing strategies by focusing more on customer budgets, their preferred services and cuisines. Apart from this it was also suggested that Zomato mustb increase its reach in more countries to increase all its stakeholders' profit.